introduce useBackNavigation hook
